refactor!: Implement cattrs converter & rename cattrs_hook to utils
